PolyesterWeave

Documentation for PolyesterWeave.

PolyesterWeave.UnsignedIteratorEarlyStop

PolyesterWeave.disable_polyester_threads

PolyesterWeave.UnsignedIteratorEarlyStop — Type

UnsignedIteratorEarlyStop(thread_mask[, num_threads = count_ones(thread_mask)])

Iterator, returning (i,t) = Tuple{UInt32,UInt32}, where i iterates from 1,2,...,num_threads, and t gives the threadids to call ThreadingUtilities.taskpointer with.

Unfortunately, codegen is suboptimal when used in the ergonomic for (i,tid) ∈ thread_iterator fashion. If you want to microoptimize, You'd get better performance from a pattern like:

function sumk(u,l = count_ones(u) % UInt32)
    uu = ServiceSolicitation.UnsignedIteratorEarlyStop(u,l)
    s = zero(UInt32); state = ServiceSolicitation.initial_state(uu)
    while true
        iter = iterate(uu, state)
        iter === nothing && break
        (i,t),state = iter
        s += t
    end
    s
end

This iterator will iterate at least once; it's important to check and exit early with a single threaded version.

PolyesterWeave.disable_polyester_threads — Method

disable_polyester_threads(f::F)

A context manager function that disables Polyester threads without affecting the scheduling of Base.Treads.@threads. Particularly useful for cases when Polyester has been used to multithread an inner small problem that is now to be used in an outer embarassingly parallel problem (in such cases it is best to multithread only at the outermost level).

This call will disable all PolyesterWeave threads, including those in LoopVectorization.@tturbo and Octavian.matmul.

Avoid calling it as @threads for i in 1:n disable_polyester_threads(f) end as that creates unnecessary per-thread overhead. Rather call it in the outermost scope, e.g. as given here:

disable_polyester_threads() do
    @threads for i in 1:n f()
end